Core: wrap sdclang commands with ccache and sync with Uber
As far as I am aware (and through testing of my own), sdclang commands are not getting wrapped for ccache.

Reason for this is in the private variable assignment section, when my_cc is stripped, it is not empty, so a wrapper is never assigned. Since it is not going to be touched until the PRIVATE_CC or PRIVATE_CXX variable assignment, we can just do it here.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>

# Conflicts:
#	core/binary.mk